Move utils around a bit
